requirements.txt
flask==2.3.3
numpy==1.24.3
matplotlib==3.7.2
Pillow==10.0.0
scipy==1.11.1
h5py==3.9.0
scikit-image==0.21.0
